Separate BackBuffer from guiutils.h
