+++ rtgui/widgets/basic/backbuffer.cc
@@ -1,0 +1,320 @@
+/*
+ *  This file is part of RawTherapee.
+ *
+ *  Copyright (c) 2004-2010 Gabor Horvath <hgabor@rawtherapee.com>
+ *
+ *  RawTherapee is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  RawTherapee is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with RawTherapee.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#include "backbuffer.h"
+
+#include "rtengine/rt_math.h"
+#include "rtengine/utils.h"
+
+#include <cassert>
+
+BackBuffer::BackBuffer() : x(0), y(0), w(0), h(0), offset(0, 0), dirty(true) {}
+BackBuffer::BackBuffer(int width, int height, Cairo::Format format) : x(0), y(0), w(width), h(height), offset(0, 0), dirty(true)
+{
+    if (w > 0 && h > 0) {
+        surface = Cairo::ImageSurface::create(format, w, h);
+    } else {
+        w = h = 0;
+    }
+}
+
+void BackBuffer::setDestPosition(int x, int y)
+{
+    // values will be clamped when used...
+    this->x = x;
+    this->y = y;
+}
+
+void BackBuffer::setSrcOffset(int x, int y)
+{
+    // values will be clamped when used...
+    offset.set(x, y);
+}
+
+void BackBuffer::setSrcOffset(const rtengine::Coord &newOffset)
+{
+    // values will be clamped when used...
+    offset = newOffset;
+}
+
+void BackBuffer::getSrcOffset(int &x, int &y)
+{
+    // values will be clamped when used...
+    offset.get(x, y);
+}
+
+void BackBuffer::getSrcOffset(rtengine::Coord &offset)
+{
+    // values will be clamped when used...
+    offset = this->offset;
+}
+
+// Note: newW & newH must be > 0
+bool BackBuffer::setDrawRectangle(Glib::RefPtr<Gdk::Window> window, Gdk::Rectangle &rectangle, bool updateBackBufferSize)
+{
+    return setDrawRectangle(window, rectangle.get_x(), rectangle.get_y(), rectangle.get_width(), rectangle.get_height(), updateBackBufferSize);
+}
+
+// Note: newW & newH must be > 0
+bool BackBuffer::setDrawRectangle(Glib::RefPtr<Gdk::Window> window, int newX, int newY, int newW, int newH, bool updateBackBufferSize)
+{
+    assert(newW && newH);
+
+    bool newSize = (newW > 0 && w != newW) || (newH > 0 && h != newH);
+
+    x = newX;
+    y = newY;
+    if (newW > 0) {
+        w = newW;
+    }
+    if (newH > 0) {
+        h = newH;
+    }
+
+    // WARNING: we're assuming that the surface type won't change during all the execution time of RT. I guess it may be wrong when the user change the gfx card display settings!?
+    if (((updateBackBufferSize && newSize) || !surface) && window) {
+        // allocate a new Surface
+        surface.clear();  // ... don't know if this is necessary?
+        surface = Cairo::ImageSurface::create(Cairo::FORMAT_RGB24, w, h);
+        dirty = true;
+    }
+
+    return dirty;
+}
+
+// Note: newW & newH must be > 0
+bool BackBuffer::setDrawRectangle(Cairo::Format format, Gdk::Rectangle &rectangle, bool updateBackBufferSize)
+{
+    return setDrawRectangle(format, rectangle.get_x(), rectangle.get_y(), rectangle.get_width(), rectangle.get_height(), updateBackBufferSize);
+}
+
+// Note: newW & newH must be > 0
+bool BackBuffer::setDrawRectangle(Cairo::Format format, int newX, int newY, int newW, int newH, bool updateBackBufferSize)
+{
+    assert(newW && newH);
+
+    bool newSize = (newW > 0 && w != newW) || (newH > 0 && h != newH);
+
+    x = newX;
+    y = newY;
+    if (newW > 0) {
+        w = newW;
+    }
+    if (newH > 0) {
+        h = newH;
+    }
+
+    // WARNING: we're assuming that the surface type won't change during all the execution time of RT. I guess it may be wrong when the user change the gfx card display settings!?
+    if ((updateBackBufferSize && newSize) || !surface) {
+        // allocate a new Surface
+        surface.clear();  // ... don't know if this is necessary?
+        surface = Cairo::ImageSurface::create(format, w, h);
+        dirty = true;
+    }
+
+    return dirty;
+}
+
+/*
+ * Copy uint8 RGB raw data to an ImageSurface. We're assuming that the source contains enough data for the given srcX, srcY, srcW, srcH -> no error checking!
+ */
+void BackBuffer::copyRGBCharData(const unsigned char *srcData, int srcX, int srcY, int srcW, int srcH, int srcRowStride, int dstX, int dstY)
+{
+    unsigned char r, g, b;
+
+    if (!surface) {
+        return;
+    }
+
+    //printf("copyRGBCharData:    src: (X:%d Y:%d, W:%d H:%d)  /  dst: (X: %d Y:%d)\n", srcX, srcY, srcW, srcH, dstX, dstY);
+
+    unsigned char *dstData = surface->get_data();
+    int surfW = surface->get_width();
+    int surfH = surface->get_height();
+
+    if (!srcData || dstX >= surfW || dstY >= surfH || srcW <= 0 || srcH <= 0 || srcX < 0 || srcY < 0) {
+        return;
+    }
+
+    for (int i = 0; i < srcH; ++i) {
+        if (dstY + i >= surfH) {
+            break;
+        }
+
+        const unsigned char *src = srcData + i * srcRowStride;
+        unsigned char *dst = dstData + ((dstY + i) * surfW + dstX) * 4;
+
+        for (int j = 0; j < srcW; ++j) {
+            if (dstX + j >= surfW) {
+                break;
+            }
+
+            r = *(src++);
+            g = *(src++);
+            b = *(src++);
+
+            rtengine::poke255_uc(dst, r, g, b);
+        }
+    }
+
+    surface->mark_dirty();
+
+}
+
+/*
+ * Copy the backbuffer to a Gdk::Window
+ */
+void BackBuffer::copySurface(Glib::RefPtr<Gdk::Window> window, Gdk::Rectangle *destRectangle)
+{
+    if (surface && window) {
+        // TODO: look out if window can be different on each call, and if not, store a reference to the window
+        Cairo::RefPtr<Cairo::Context> crSrc = window->create_cairo_context();
+        Cairo::RefPtr<Cairo::Surface> destSurface = crSrc->get_target();
+
+        // compute the source offset
+        int offsetX = rtengine::LIM<int>(offset.x, 0, surface->get_width());
+        int offsetY = rtengine::LIM<int>(offset.y, 0, surface->get_height());
+
+        // now copy the off-screen Surface to the destination Surface
+        Cairo::RefPtr<Cairo::Context> crDest = Cairo::Context::create(destSurface);
+        crDest->set_line_width(0.);
+
+        if (destRectangle) {
+            crDest->set_source(surface, -offsetX + destRectangle->get_x(), -offsetY + destRectangle->get_y());
+            int w_ = destRectangle->get_width() > 0 ? destRectangle->get_width() : w;
+            int h_ = destRectangle->get_height() > 0 ? destRectangle->get_height() : h;
+            //printf("BackBuffer::copySurface / rectangle1(%d, %d, %d, %d)\n", destRectangle->get_x(), destRectangle->get_y(), w_, h_);
+            crDest->rectangle(destRectangle->get_x(), destRectangle->get_y(), w_, h_);
+            //printf("BackBuffer::copySurface / rectangle1\n");
+        } else {
+            crDest->set_source(surface, -offsetX + x, -offsetY + y);
+            //printf("BackBuffer::copySurface / rectangle2(%d, %d, %d, %d)\n", x, y, w, h);
+            crDest->rectangle(x, y, w, h);
+            //printf("BackBuffer::copySurface / rectangle2\n");
+        }
+
+        crDest->fill();
+    }
+}
+
+/*
+ * Copy the BackBuffer to another BackBuffer
+ */
+void BackBuffer::copySurface(BackBuffer *destBackBuffer, Gdk::Rectangle *destRectangle)
+{
+    if (surface && destBackBuffer) {
+        Cairo::RefPtr<Cairo::ImageSurface> destSurface = destBackBuffer->getSurface();
+
+        if (!destSurface) {
+            return;
+        }
+
+        // compute the source offset
+        int offsetX = rtengine::LIM<int>(offset.x, 0, surface->get_width());
+        int offsetY = rtengine::LIM<int>(offset.y, 0, surface->get_height());
+
+        // now copy the off-screen Surface to the destination Surface
+        Cairo::RefPtr<Cairo::Context> crDest = Cairo::Context::create(destSurface);
+        crDest->set_line_width(0.);
+
+        if (destRectangle) {
+            crDest->set_source(surface, -offsetX + destRectangle->get_x(), -offsetY + destRectangle->get_y());
+            int w_ = destRectangle->get_width() > 0 ? destRectangle->get_width() : w;
+            int h_ = destRectangle->get_height() > 0 ? destRectangle->get_height() : h;
+            //printf("BackBuffer::copySurface / rectangle3(%d, %d, %d, %d)\n", destRectangle->get_x(), destRectangle->get_y(), w_, h_);
+            crDest->rectangle(destRectangle->get_x(), destRectangle->get_y(), w_, h_);
+            //printf("BackBuffer::copySurface / rectangle3\n");
+        } else {
+            crDest->set_source(surface, -offsetX + x, -offsetY + y);
+            //printf("BackBuffer::copySurface / rectangle4(%d, %d, %d, %d)\n", x, y, w, h);
+            crDest->rectangle(x, y, w, h);
+            //printf("BackBuffer::copySurface / rectangle4\n");
+        }
+
+        crDest->fill();
+    }
+}
+
+/*
+ * Copy the BackBuffer to another Cairo::Surface
+ */
+void BackBuffer::copySurface(const Cairo::RefPtr<Cairo::ImageSurface>& destSurface,
+                             Gdk::Rectangle *destRectangle)
+{
+    if (surface && destSurface) {
+        // compute the source offset
+        int offsetX = rtengine::LIM<int>(offset.x, 0, surface->get_width());
+        int offsetY = rtengine::LIM<int>(offset.y, 0, surface->get_height());
+
+        // now copy the off-screen Surface to the destination Surface
+        Cairo::RefPtr<Cairo::Context> crDest = Cairo::Context::create(destSurface);
+        crDest->set_line_width(0.);
+
+        if (destRectangle) {
+            crDest->set_source(surface, -offsetX + destRectangle->get_x(), -offsetY + destRectangle->get_y());
+            int w_ = destRectangle->get_width() > 0 ? destRectangle->get_width() : w;
+            int h_ = destRectangle->get_height() > 0 ? destRectangle->get_height() : h;
+            //printf("BackBuffer::copySurface / rectangle5(%d, %d, %d, %d)\n", destRectangle->get_x(), destRectangle->get_y(), w_, h_);
+            crDest->rectangle(destRectangle->get_x(), destRectangle->get_y(), w_, h_);
+            //printf("BackBuffer::copySurface / rectangle5\n");
+        } else {
+            crDest->set_source(surface, -offsetX + x, -offsetY + y);
+            //printf("BackBuffer::copySurface / rectangle6(%d, %d, %d, %d)\n", x, y, w, h);
+            crDest->rectangle(x, y, w, h);
+            //printf("BackBuffer::copySurface / rectangle6\n");
+        }
+
+        crDest->fill();
+    }
+}
+
+/*
+ * Copy the BackBuffer to another Cairo::Surface
+ */
+void BackBuffer::copySurface(const Cairo::RefPtr<Cairo::Context>& crDest,
+                             Gdk::Rectangle *destRectangle)
+{
+    if (surface && crDest) {
+        // compute the source offset
+        int offsetX = rtengine::LIM<int>(offset.x, 0, surface->get_width());
+        int offsetY = rtengine::LIM<int>(offset.y, 0, surface->get_height());
+
+        // now copy the off-screen Surface to the destination Surface
+        // int srcSurfW = surface->get_width();
+        // int srcSurfH = surface->get_height();
+        //printf("srcSurf:  w: %d, h: %d\n", srcSurfW, srcSurfH);
+        crDest->set_line_width(0.);
+
+        if (destRectangle) {
+            crDest->set_source(surface, -offsetX + destRectangle->get_x(), -offsetY + destRectangle->get_y());
+            int w_ = destRectangle->get_width() > 0 ? destRectangle->get_width() : w;
+            int h_ = destRectangle->get_height() > 0 ? destRectangle->get_height() : h;
+            //printf("BackBuffer::copySurface / rectangle7(%d, %d, %d, %d)\n", destRectangle->get_x(), destRectangle->get_y(), w_, h_);
+            crDest->rectangle(destRectangle->get_x(), destRectangle->get_y(), w_, h_);
+            //printf("BackBuffer::copySurface / rectangle7\n");
+        } else {
+            crDest->set_source(surface, -offsetX + x, -offsetY + y);
+            //printf("BackBuffer::copySurface / rectangle8(%d, %d, %d, %d)\n", x, y, w, h);
+            crDest->rectangle(x, y, w, h);
+            //printf("BackBuffer::copySurface / rectangle8\n");
+        }
+
+        crDest->fill();
+    }
+}
